+++ src/CompilerRuntime.php
@@ -14,7 +14,7 @@ namespace JmesPath;
  */
 class CompilerRuntime
 {
-    const CACHE_VERSION = 2;
+    const CACHE_VERSION = 3;
 
     private $parser;
     private $compiler;